Since last commit: 	lost all previous commits 	now have hide functions of hint, check & solve buttons 	UIdownPicker still crashes app when trying to enter chosen value

+++ SudokuApp/MainMenuViewController.h
@@ -12,7 +12,7 @@
 
 @property (weak, nonatomic) IBOutlet UILabel *SudokuApp;
 @property (weak, nonatomic) IBOutlet UIPickerView *GameModePicker;
-@property NSInteger mode;
+@property int mode;
 
 
 - (IBAction)Go:(UIButton *)sender;
